refactor(titlebar): extract hover-gate state machine into useTitleBarHoverGate
Amp-Thread-ID: https://ampcode.com/threads/T-019e23b2-13b0-726c-9d4f-df64a04c5be7
