+++ docs/CLEANUP_PLAN.md
@@ -207,7 +207,7 @@ These are concrete hot-spots that will hurt when the feature lands. Address only
 | File | Action | Notes |
 |---|---|---|
 | `PanelApp.vue` | ✅ **Done** | Per-panel routing → composables. Done: `useAppUpdatePrompts`, `useSendFeedback`, `useDeepLinkRouter`, [`usePanelOverlays`](../src/renderer/src/panel/usePanelOverlays.ts) (overlay slot + Tier 2/3 helpers — `handleShowProgress`, `openFlowTakeover`, `openFirstUseTakeover`, `dismissTakeoverDirect`, `switchPanel`), [`useChooserHandoff`](../src/renderer/src/panel/useChooserHandoff.ts) (install-less chooser-tile launch + in-place attach claim + close-on-instance-started fallback), [`useFirstUseChain`](../src/renderer/src/panel/useFirstUseChain.ts) (cloud / chain-local / chain-migrate handlers + auto-launch watcher). PanelApp.vue 1203 → 456 lines. The three composables form a small dependency cycle (each consumes helpers from the others); broken with deferred lazy callbacks at construction time. |
-| `TitleBarApp.vue` | 🟡 **Split (in progress)** | Pull pill buttons + tooltip coordinator + update-pill state into composables. The tooltip composable will pair with **P1**. Done so far: `useTitleBarTooltip`, `useUpdatePills`, `useTitleBarMenus` (file-menu + downloads-tray + menu-reopen-guard map + isMenuOpen + downloadsState), `useTitleBarIdentity` (installLabel + sourceCategory + theme + fullscreen + first-use mode + install-type icon computeds + isLight). 1090 → 730 lines. Remaining shell carries: hover-gate state machine, install-less / activePanel refs, send-feedback handler, the wiring between the composables, and the template. |
+| `TitleBarApp.vue` | 🟡 **Split (in progress)** | Pull pill buttons + tooltip coordinator + update-pill state into composables. The tooltip composable will pair with **P1**. Done so far: `useTitleBarTooltip`, `useUpdatePills`, `useTitleBarMenus` (file-menu + downloads-tray + menu-reopen-guard map + isMenuOpen + downloadsState), `useTitleBarIdentity` (installLabel + sourceCategory + theme + fullscreen + first-use mode + install-type icon computeds + isLight), [`useTitleBarHoverGate`](../src/renderer/src/comfyTitleBar/useTitleBarHoverGate.ts) (`isHoverActive` + window-blur / pointermove / pointerleave listeners — drives `:hover` gating across menu, nav, and pill buttons; covered by [`e2e/title-bar-hover-gate.test.ts`](../e2e/title-bar-hover-gate.test.ts)). 1090 → 674 lines. Remaining shell carries: install-less / activePanel refs, send-feedback handler, the wiring between the composables, and the template. |
 | `SnapshotTab.vue` | ✅ **Done** | List / inspector are split. Inspector body (diff toggle + environment grid + custom-node + pip-package lists) extracted to [`SnapshotInspector.vue`](../src/renderer/src/components/SnapshotInspector.vue). Pure presentation helpers (`triggerClass`, `formatRelative`, `copyReasonLabel`, `changeSummary`, `diffHasChanges`) promoted to [`lib/snapshots.ts`](../src/renderer/src/lib/snapshots.ts) so the parent and inspector share one source. SnapshotTab.vue 1011 → 694 lines. |
 | `ChooserView.vue` | ✅ **Done** | Per-install tile rendering (icon, kebab/error cluster, progress block, pills, close-instance CTA) extracted to [`chooser/ChooserInstallTile.vue`](../src/renderer/src/views/chooser/ChooserInstallTile.vue). The `chooser-tile-*` CSS lives in [`chooser/chooser-tiles.css`](../src/renderer/src/views/chooser/chooser-tiles.css) and is imported from both component scoped style blocks so the New Install / Cloud entry tiles and per-install tiles render identically without duplicating the rules. ChooserView.vue 924 → 306 lines. |
 | `comfyDownloadManager.ts` | ✅ **Done** | IPC registration extracted to [`lib/ipc/registerDownloadHandlers.ts`](../src/main/lib/ipc/registerDownloadHandlers.ts). |
@@ -225,7 +225,7 @@ After the file splits, do one pass over each new module to apply AGENTS.md rules
 - Collapse 10–25-line jsdoc blocks on simple helpers to one sentence.
 - Keep the genuinely useful "why" comments (race conditions, OS quirks, IPC ordering invariants).
 
-Sweep status (Thread F): scanning the new modules in `src/main/host/*`, `src/main/popups/*`, the renderer composables (`useAppUpdatePrompts`, `useSendFeedback`, `useDeepLinkRouter`, `useTitleBarMenus`, `useTitleBarIdentity`, `useTitleBarTooltip`, `useUpdatePills`, `usePanelOverlays`, `useChooserHandoff`, `useFirstUseChain`), and `src/main/lib/ipc/registerDownloadHandlers.ts` for AGENTS.md violations turned up no historical-narration comments and a single `P0.9` plan reference (in `host/detach.ts`, fixed). The remaining long doc-blocks are genuine "why" notes (registry late-binding rationale, popup lifecycle invariants, IPC ordering, overlay tier-collision rules, first-use chain auto-launch invariant) that AGENTS.md tells us to keep. No further trimming required for these modules.
+Sweep status (Thread F): scanning the new modules in `src/main/host/*`, `src/main/popups/*`, the renderer composables (`useAppUpdatePrompts`, `useSendFeedback`, `useDeepLinkRouter`, `useTitleBarMenus`, `useTitleBarIdentity`, `useTitleBarTooltip`, `useUpdatePills`, `useTitleBarHoverGate`, `usePanelOverlays`, `useChooserHandoff`, `useFirstUseChain`), and `src/main/lib/ipc/registerDownloadHandlers.ts` for AGENTS.md violations turned up no historical-narration comments and a single `P0.9` plan reference (in `host/detach.ts`, fixed). The remaining long doc-blocks are genuine "why" notes (registry late-binding rationale, popup lifecycle invariants, IPC ordering, overlay tier-collision rules, first-use chain auto-launch invariant) that AGENTS.md tells us to keep. No further trimming required for these modules.
 
 ---
